# src/config/swagger/api.schema.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WoTTranslator",
    "description": "The WoTTranslator converts OpenAPI schema to W3C WoT Thing Descriptions and deploys them",
    "contact": {
      "email": "ivandimitry.ribeiro@unibo.it",
      "name": "Ivan Zyrianoff"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3334"
    }
  ],
  "paths": {
    "/deployWoT": {
      "post": {
        "summary": "deploys a Web Thing from a OpenAPI schema",
        "description": "Translated a OpenAPI Schema into a WoT Thing Description and deploy the correspondent Web Thing",
        "responses": {
          "201": {
            "description": "successful operation"
          },
          "503": {
            "description": "Service Unavailable"
          },
          "409": {
            "description": "There is already a Web Thing deployed with the same title"
          },
          "400": {
            "description": "Invalid Request"
          }
        }
      }
    },
    "/deployWoT/url": {
      "post": {
        "summary": "deploys a Web Thing given a URL with a OpenAPI schema",
        "description": "Perform a GET request in the provided url that must contain a OpenAPI schema. Then, the application translates the obtained OpenAPI Schema into a WoT Thing Description and deploy the correspondent Web Thing",
        "responses": {
          "201": {
            "description": "successful operation"
          },
          "503": {
            "description": "Service Unavailable"
          },
          "409": {
            "description": "There is already a Web Thing deployed with the same title"
          },
          "400": {
            "description": "Invalid Request"
          },
          "404": {
            "description": "Error fetching the requested URL"
          }
        }
      }
    },
    "/translateOpenApi": {
      "post": {
        "summary": "Returns a Thing Description",
        "description": "Translated a OpenAPI Schema into a WoT Thing Description",
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "503": {
            "description": "Service Unavailable"
          },
          "400": {
            "description": "Invalid Request"
          }
        }
      }
    },
    "/translateOpenApi/url": {
      "post": {
        "summary": "Returns a Thing Description",
        "description": "Perform a GET request in the provided url that must contain a OpenAPI schema. Then, the application translates the obtainedTranslated a OpenAPI Schema into a WoT Thing Description",
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "503": {
            "description": "Service Unavailable"
          },
          "400": {
            "description": "Invalid Request"
          }
        }
      }
    },
    "/health": {
      "get": {
        "description": "returns an object containing application metadata if the application is current running",
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    }
  }
}